pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfy-bridge"
version = "0.1.0"
description = "ComfyUI Bridge for AI Power Grid"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "httpx",
  "aiohttp",
  "python-dotenv",
  "pytest",
  "pytest-asyncio"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bridge*"]